ProjectImplementation/usecase1/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jewelry-ai"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    # Configuration
    "pydantic-settings>=2.5",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pgvector>=0.3",
    # Cache + task queue
    "redis[hiredis]>=5.0",
    "celery[redis]>=5.4",
    # HTTP client
    "httpx>=0.27",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # AI / LLM
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "langgraph>=0.2",
    "openai>=1.50",
    # ML
    "xgboost>=2.1",
    "scikit-learn>=1.5",
    "pandas>=2.2",
    "numpy>=1.26",
    "mlflow>=2.16",
    # Integrations
    "sendgrid>=6.11",
    "twilio>=9.3",
    # UI
    "streamlit>=1.40",
    "plotly>=5.24",
    # Observability
    "prometheus-fastapi-instrumentator>=7.0",
    "slowapi>=0.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "factory-boy>=3.3",
    "respx>=0.21",
    "freezegun>=1.5",
    "Faker>=30.0",
    # Code quality
    "ruff>=0.7",
    "mypy>=1.13",
    # Type stubs
    "types-passlib>=1.7",
    "types-python-jose>=3.3",
    "pandas-stubs>=2.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ── Ruff ─────────────────────────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = [
    "B008",   # FastAPI Depends() pattern — do not call in argument defaults
    "N818",   # Exception names — some intentionally don't end in Error
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "N806",   # PascalCase mock variable names are conventional in tests (MockFoo, MockRepo)
    "E501",   # Long lines acceptable in test fixtures and inline data
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ── Mypy ─────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/"]

[[tool.mypy.overrides]]
module = ["celery.*", "xgboost.*", "mlflow.*", "streamlit.*", "plotly.*", "sendgrid.*", "twilio.*", "langchain.*", "langgraph.*"]
ignore_missing_imports = true

# ── Pytest ───────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: unit tests (fast, all dependencies mocked)",
    "integration: integration tests (requires running PostgreSQL and Redis)",
    "e2e: end-to-end tests (full stack — Docker Compose must be running)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ── Coverage ─────────────────────────────────────────────────────────────────

[tool.coverage.run]
source = ["src"]
omit = [
    "src/main.py",
    "*/__init__.py",
    "src/scrapers/*",
    "src/ml/fine_tuning/*",
]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@overload",
    "if __name__ == .__main__.:",
]
